application/requirements.txt
wheel
amqp==2.5.2
asgi-redis==1.4.3
asgiref==1.1.2
attrs==19.3.0
autobahn==20.12.3
Automat==0.8.0
billiard==3.6.1.0
celery==5.2.2
certifi==2019.11.28
cffi==1.13.2
channels==1.1.3
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
cssselect==1.1.0
d==0.2.2
daphne==1.4.1
django~=2.2
django-adminplus==0.5
django-channels-presence==0.0.9
django-chunked-upload==2.0.0
django-ckeditor==5.8.0
django-cors-headers==3.1.1
django-filter==2.4.0
django-hexdi==0.2.1
django-jenkins==0.110.0
django-jet==1.0.8
django-js-asset==1.2.2
django-nested-admin==3.2.4
django-tinymce==3.0.2
djangorestframework==3.11.2
dominate==2.4.0
drf-yasg==1.17.0
gevent==1.5a4
greenlet==0.4.15
hexdi==0.2.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.3
kombu==4.6.6
lxml==4.9.1
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack-python==0.5.6
mysqlclient==1.4.5
napkin==0.5.4.9
packaging==20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.7.4
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyquery==1.4.1
python-magic==0.4.15
python-monkey-business==1.0.0
pytz==2019.3
redis==2.10.6
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
Twisted==22.10.0
txaio==18.8.1
uritemplate==3.0.1
vine==1.3.0
zipp==0.6.0
PyInstaller==3.6
urllib3
inject
requests
django-fullurl
django_extensions
